FEATURE( delete account ): Delete development.

+++ accounts/src/main/java/accounts/services/AccountsDeleteRedisListenerService.java
@@ -1,13 +1,19 @@
 package accounts.services;
 
+import accounts.persistence.entities.AccountsAddressEntity;
+import accounts.persistence.entities.AccountsDeletedEntity;
 import accounts.persistence.entities.AccountsEntity;
 import accounts.persistence.entities.AccountsProfileEntity;
+import accounts.persistence.repositories.AccountsDeletedRepository;
 import accounts.persistence.repositories.AccountsProfileRepository;
 import accounts.persistence.repositories.AccountsRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.data.redis.connection.Message;
 import org.springframework.data.redis.connection.MessageListener;
 import org.springframework.stereotype.Component;
 
+import java.time.ZoneOffset;
+import java.time.ZonedDateTime;
 import java.util.Optional;
 import java.util.UUID;
 
@@ -16,21 +22,31 @@ public class AccountsDeleteRedisListenerService implements MessageListener {
 
     private final AccountsRepository accountsRepository;
     private final AccountsProfileRepository accountsProfileRepository;
+    private final AccountsManagementService accountsManagementService;
+    private final EncryptionService encryptionService;
+    private final AccountsDeletedRepository accountsDeletedRepository;
 
     public AccountsDeleteRedisListenerService (
 
         AccountsRepository accountsRepository,
-        AccountsProfileRepository accountsProfileRepository
+        AccountsProfileRepository accountsProfileRepository,
+        AccountsManagementService accountsManagementService,
+        EncryptionService encryptionService,
+        AccountsDeletedRepository accountsDeletedRepository
 
     ) {
 
         this.accountsRepository = accountsRepository;
         this.accountsProfileRepository = accountsProfileRepository;
+        this.accountsManagementService = accountsManagementService;
+        this.encryptionService = encryptionService;
+        this.accountsDeletedRepository = accountsDeletedRepository;
 
     }
 
     // Delete account (user decision)
     @Override
+    @Transactional
     public void onMessage(Message message, byte[] pattern) {
 
         String expiredKey  = new String(message.getBody());
@@ -39,7 +55,28 @@ public class AccountsDeleteRedisListenerService implements MessageListener {
 
         if (expiredKey.startsWith("deletedAccountByUserCache")) {
 
-            System.out.println(expiredKey);
+            // Get user account
+            Optional<AccountsEntity> findUser =  accountsRepository.findById(
+                idUser
+            );
+
+            if (findUser.isPresent()) {
+
+                // Change email for user id
+                findUser.get().setEmail("deleted-" + findUser.get().getId().toString());
+                findUser.get().setPassword(
+                    encryptionService.hashPassword(UUID.randomUUID().toString())
+                );
+                accountsRepository.save(findUser.get());
+
+                // Create deleted account (id user for id and email)
+                AccountsDeletedEntity newDeletedAccount = new AccountsDeletedEntity();
+                newDeletedAccount.setCreatedAt(ZonedDateTime.now(ZoneOffset.UTC).toInstant());
+                newDeletedAccount.setId(findUser.get().getId());
+                newDeletedAccount.setEmail(findUser.get().getEmail());
+                accountsDeletedRepository.save(newDeletedAccount);
+
+            }
 
         }
 
@@ -55,9 +92,13 @@ public class AccountsDeleteRedisListenerService implements MessageListener {
             Optional<AccountsProfileEntity> findProfile = accountsProfileRepository
                 .findById(idUser);
 
-            // Delete both
-            accountsRepository.delete(findUser.get());
-            accountsProfileRepository.delete(findProfile.get());
+            if ( findUser.isPresent() && findProfile.isPresent() ) {
+
+                // Delete both
+                accountsRepository.delete(findUser.get());
+                accountsProfileRepository.delete(findProfile.get());
+
+            }
 
         }
